Decoupling AttackOrb
The AttackOrb is now to be a new object inherited from a BaseAttack. The BaseAttack objects are contained within an AttackContainer which simply manages the movements of the attacks, the BaseAttack objects will then implement the logic of the live attack
