+++ Assets/Resources/AttackOrb.prefab
@@ -92,7 +92,7 @@ Animator:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 3573619596765284886}
-  m_Enabled: 1
+  m_Enabled: 0
   m_Avatar: {fileID: 0}
   m_Controller: {fileID: 9100000, guid: 2b77debf1905a8f4aafbec7d14a54f2d, type: 2}
   m_CullingMode: 0
@@ -209,7 +209,7 @@ GameObject:
   m_Component:
   - component: {fileID: 4441261147921284305}
   - component: {fileID: 551306284263082038}
-  - component: {fileID: 8343826703721570159}
+  - component: {fileID: 2761361259800092659}
   m_Layer: 14
   m_Name: AttackOrb
   m_TagString: Untagged
@@ -338,7 +338,7 @@ TrailRenderer:
   m_MinVertexDistance: 0.1
   m_Autodestruct: 0
   m_Emitting: 1
---- !u!114 &8343826703721570159
+--- !u!114 &2761361259800092659
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -347,23 +347,20 @@ MonoBehaviour:
   m_GameObject: {fileID: 4441261147921284307}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8ba99ecb435bc574084b778c3b9c51c2, type: 3}
+  m_Script: {fileID: 11500000, guid: b371897cb307d114e8fe0777b8d52214, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  rotateFrequency: 1.5
-  strikeForce: 5
-  thisFacingDirection: 2
-  attacking: 0
-  finalFlingDirection: {x: 0, y: -10.5}
-  xOffset: 0.25
-  yOffset: 0.25
+  chargeLimit: 2
+  velocityModifier: 0.1
+  enabledTime: 1
+  impactForce: 10
+  chargeUpSound: {fileID: 0}
+  chargedFullySound: {fileID: 0}
+  attackReleaseSound: {fileID: 0}
   lifeTimeCounter: 0
-  lifeTimeLimit: 5
-  distancefactor: 9
-  speed: 0.14
-  chargeUpSound: {fileID: 8300000, guid: ba9407ab328f0b74db8205d4d9d19ec4, type: 3}
-  chargedFullySound: {fileID: 8300000, guid: 2f846ce7b011b4447b7164620357d799, type: 3}
-  attackReleaseSound: {fileID: 8300000, guid: 8c0ce1455ba838c498ca6fd0a1cc8e3b, type: 3}
+  lifeTimeLimit: 0
+  distancefactor: 5
+  speed: 0
   damage: 1
 --- !u!1 &4953624384292263635
 GameObject:
@@ -403,7 +400,7 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 4953624384292263635}
-  m_Enabled: 0
+  m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 1db55a2b76e5f0f44a76181ed1a1ae60, type: 3}
   m_Name:
