+++ Assets/Resources/Player.prefab
@@ -14,7 +14,6 @@ GameObject:
   - component: {fileID: 2430776021576682865}
   - component: {fileID: 2430776021576682864}
   - component: {fileID: 707757449049624803}
-  - component: {fileID: 4180652227649236425}
   - component: {fileID: 4514698623749904686}
   - component: {fileID: 1079849769528210242}
   - component: {fileID: 1077947487583757974}
@@ -38,8 +37,8 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 2430776022666849404}
-  - {fileID: 5669393226998340586}
   - {fileID: 4307181509312280477}
+  - {fileID: 425641250105451400}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -116,18 +115,6 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 4ec2d2955bc258a4bb42216fa556feda, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!114 &4180652227649236425
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2430776021576682866}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: eec6924da51f96a43a1d67766d436327, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!114 &4514698623749904686
@@ -484,7 +471,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 2430776021576682879}
-  m_RootOrder: 2
+  m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!50 &1529188420793098683
 Rigidbody2D:
@@ -595,13 +582,125 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!1001 &8290911829000628027
+--- !u!1 &4851514473392383607
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 425641250105451400}
+  - component: {fileID: 6425308278114801946}
+  - component: {fileID: 6176650463867773506}
+  m_Layer: 0
+  m_Name: AttackContainer
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &425641250105451400
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4851514473392383607}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children:
+  - {fileID: 1226646838991694219}
+  m_Father: {fileID: 2430776021576682879}
+  m_RootOrder: 2
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &6425308278114801946
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4851514473392383607}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: eec6924da51f96a43a1d67766d436327, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  cooldown: 2
+  rotateFrequency: 5
+--- !u!212 &6176650463867773506
+SpriteRenderer:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4851514473392383607}
+  m_Enabled: 0
+  m_CastShadows: 0
+  m_ReceiveShadows: 0
+  m_DynamicOccludee: 1
+  m_MotionVectors: 1
+  m_LightProbeUsage: 1
+  m_ReflectionProbeUsage: 1
+  m_RayTracingMode: 0
+  m_RenderingLayerMask: 1
+  m_RendererPriority: 0
+  m_Materials:
+  - {fileID: 10754, guid: 0000000000000000f000000000000000, type: 0}
+  m_StaticBatchInfo:
+    firstSubMesh: 0
+    subMeshCount: 0
+  m_StaticBatchRoot: {fileID: 0}
+  m_ProbeAnchor: {fileID: 0}
+  m_LightProbeVolumeOverride: {fileID: 0}
+  m_ScaleInLightmap: 1
+  m_ReceiveGI: 1
+  m_PreserveUVs: 0
+  m_IgnoreNormalsForChartDetection: 0
+  m_ImportantGI: 0
+  m_StitchLightmapSeams: 1
+  m_SelectedEditorRenderState: 0
+  m_MinimumChartSize: 4
+  m_AutoUVMaxDistance: 0.5
+  m_AutoUVMaxAngle: 89
+  m_LightmapParameters: {fileID: 0}
+  m_SortingLayerID: 551469179
+  m_SortingLayer: 2
+  m_SortingOrder: 5
+  m_Sprite: {fileID: 21300000, guid: d31ef7e4c318b2240889c83740943596, type: 3}
+  m_Color: {r: 1, g: 1, b: 1, a: 1}
+  m_FlipX: 0
+  m_FlipY: 0
+  m_DrawMode: 0
+  m_Size: {x: 0.15, y: 0.16}
+  m_AdaptiveModeThreshold: 0.5
+  m_SpriteTileMode: 0
+  m_WasSpriteAssigned: 1
+  m_MaskInteraction: 0
+  m_SpriteSortPoint: 0
+--- !u!1001 &3217658891334833498
 PrefabInstance:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Modification:
-    m_TransformParent: {fileID: 2430776021576682879}
+    m_TransformParent: {fileID: 425641250105451400}
     m_Modifications:
+    - target: {fileID: 2761361259800092659, guid: 8c7e04daea270274480cd156167d57b0,
+        type: 3}
+      propertyPath: impactForce
+      value: 20
+      objectReference: {fileID: 0}
+    - target: {fileID: 2761361259800092659, guid: 8c7e04daea270274480cd156167d57b0,
+        type: 3}
+      propertyPath: enabledTime
+      value: 0.66
+      objectReference: {fileID: 0}
+    - target: {fileID: 2761361259800092659, guid: 8c7e04daea270274480cd156167d57b0,
+        type: 3}
+      propertyPath: velocityModifier
+      value: 0.224
+      objectReference: {fileID: 0}
     - target: {fileID: 4441261147921284305, guid: 8c7e04daea270274480cd156167d57b0,
         type: 3}
       propertyPath: m_LocalPosition.x
@@ -640,7 +739,7 @@ PrefabInstance:
     - target: {fileID: 4441261147921284305, guid: 8c7e04daea270274480cd156167d57b0,
         type: 3}
       propertyPath: m_RootOrder
-      value: 1
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 4441261147921284305, guid: 8c7e04daea270274480cd156167d57b0,
         type: 3}
@@ -662,16 +761,11 @@ PrefabInstance:
       propertyPath: m_Name
       value: AttackOrb
       objectReference: {fileID: 0}
-    - target: {fileID: 6884574825524571333, guid: 8c7e04daea270274480cd156167d57b0,
-        type: 3}
-      propertyPath: m_BodyType
-      value: 0
-      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 8c7e04daea270274480cd156167d57b0, type: 3}
---- !u!4 &5669393226998340586 stripped
+--- !u!4 &1226646838991694219 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 4441261147921284305, guid: 8c7e04daea270274480cd156167d57b0,
     type: 3}
-  m_PrefabInstance: {fileID: 8290911829000628027}
+  m_PrefabInstance: {fileID: 3217658891334833498}
   m_PrefabAsset: {fileID: 0}
