+++ Assets/Scripts/Player/Strikes/BaseAttackObject.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e9673249166b1d64ea7fdf58f3805413
+guid: b371897cb307d114e8fe0777b8d52214
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
