+++ Assets/Scripts/Player/Strikes/oldshit.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: dbef8127ec1fb4748b5de86385c759eb
+guid: fcf74e11c8d5f5443a997ec7299bb72c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
